ai_powered_legal_council/requirements.txt
streamlit
google-generativeai
pymupdf
pytesseract
pillow
